Move buildTemplateArgs to the template_wrapper module
