+++ js/renovation/component_wrapper/common/component.ts
@@ -16,7 +16,7 @@ import type { UserDefinedElement } from '../../../core/element';
 import {
   isDefined, isRenderer, isString,
 } from '../../../core/utils/type';
-import { TemplateModel, TemplateWrapper, TemplateWrapperProps } from './template_wrapper';
+import { TemplateModel, TemplateWrapper, buildTemplateArgs } from './template_wrapper';
 import { updatePropsImmutable } from '../utils/update_props_immutable';
 import type { Option, TemplateComponent } from './types';
 
@@ -33,12 +33,6 @@ const setDefaultOptionValue = (
 interface ElementAttributes extends Record<string, unknown> {
   class: string;
 }
-
-type TemplateModelArgs =
-  // eslint-disable-next-line @typescript-eslint/no-type-alias
-  Required<Pick<TemplateWrapperProps, 'model'>>
-  // eslint-disable-next-line @typescript-eslint/no-type-alias
-  & Omit<TemplateWrapperProps, 'model'>;
 
 export interface ComponentWrapperProps extends Record<string, unknown> {
   onContentReady?: (e: Record<string, unknown>) => void;
@@ -488,24 +482,6 @@ export default class ComponentWrapper extends DOMComponent<ComponentWrapperProps
     return null;
   }
 
-  _buildTemplateArgs(
-    model: TemplateModel,
-    template: TemplateWrapperProps['template'],
-  ): TemplateModelArgs {
-    const args: TemplateModelArgs = {
-      template,
-      model: { ...model },
-    };
-
-    const { isEqual, ...data } = model.data ?? {};
-    if (isEqual) {
-      args.model.data = data;
-      args.isEqual = isEqual;
-    }
-
-    return args;
-  }
-
   _createTemplateComponent(templateOption: unknown): TemplateComponent | undefined {
     if (!templateOption) {
       return undefined;
@@ -516,8 +492,6 @@ export default class ComponentWrapper extends DOMComponent<ComponentWrapperProps
     if (isString(template) && template === 'dx-renovation-template-mock') {
       return undefined;
     }
-
-    const buildTemplateArgs = this._buildTemplateArgs;
 
     // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     const templateWrapper = (model: TemplateModel): VNode => renderer.createElement(
